misc: therm: therm_fan_est: Convert temp to integer for removing warning

bug 200261578

Reviewed-on: http://git-master/r/1299268

+++ include/linux/therm_est.h
@@ -98,7 +98,7 @@ struct therm_est_data {
 
 struct therm_fan_est_subdevice {
 	const char *dev_data;
-	int (*get_temp)(const char *, long *);
+	int (*get_temp)(const char *, int *);
 	int group;
 	/*
 	 * as we read coeffs array from the device tree,
